pyproject.toml
[project]
name = "finagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-secretsmanager-caching>=1.1.3",
    "azure-monitor-opentelemetry>=1.6.7",
    "azure-storage-blob>=12.28.0",
    "boto3>=1.42.49",
    "botocore>=1.42.49",
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "fastapi>=0.129.0",
    "firecrawl-py>=4.14.1",
    "openai>=1.83.0",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "s3fs>=2026.2.0",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
    "watchtower>=3.4.0",
    "yfinance>=1.1.0",
]
